+++ packages/sampler/stories/moonstone-stories/Button.js
@@ -1,4 +1,4 @@
-import Button, {ButtonBase} from '@enact/moonstone/Button';
+import Button, {ButtonBaseFactory} from '@enact/moonstone/Button';
 import {icons} from '@enact/moonstone/Icon';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
@@ -7,7 +7,7 @@ import {boolean, select, text} from '@kadira/storybook-addon-knobs';
 import {mergeComponentMetadata} from '../../src/utils/propTables';
 import nullify from '../../src/utils/nullify.js';
 
-const Config = mergeComponentMetadata('Button', ButtonBase, Button);
+const Config = mergeComponentMetadata('Button', ButtonBaseFactory(), Button);
 
 // Set up some defaults for info and knobs
 const prop = {
@@ -24,11 +24,11 @@ storiesOf('Button')
 				onClick={action('onClick')}
 				backgroundOpacity={nullify(select('backgroundOpacity', prop.backgroundOpacity))}
 				casing={select('casing', ['preserve', 'sentence', 'word', 'upper'], 'upper')}
-				disabled={boolean('disabled', ButtonBase.defaultProps.disabled)}
-				icon={nullify(select('icon', prop.icons))}
-				minWidth={nullify(boolean('minWidth', ButtonBase.defaultProps.minWidth))}
+				disabled={boolean('disabled', Config.defaultProps.disabled)}
+				icon={nullify(select('icon', prop.icons, Config.defaultProps.icon))}
+				minWidth={nullify(boolean('minWidth', Config.defaultProps.minWidth))}
 				selected={nullify(boolean('selected', false))}
-				small={nullify(boolean('small', ButtonBase.defaultProps.small))}
+				small={nullify(boolean('small', Config.defaultProps.small))}
 			>
 				{text('children', 'click me')}
 			</Button>